requirements.txt
azure-identity==1.19.0
azure-monitor-opentelemetry==1.6.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
azure-ai-projects==1.0.0b5
azure-ai-inference==1.0.0b7
aiohttp==3.11.11
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyGithub==2.5.0
